rename CameraHomeButton file to match new default export
