+++ src/components/PlaybackControls/index.tsx
@@ -12,9 +12,9 @@ import { IconGlyphs } from "../../constants/interfaces";
 import { getIconGlyphClasses } from "../../util";
 import { Pause, Play } from "../Icons";
 import ViewportButton from "../ViewportButton";
+import CameraHomeButton from "../CameraControls/CameraHomeButton";
 
 import styles from "./style.css";
-import { CameraHomeButton } from "../CameraControls/HomeButton";
 
 interface PlayBackProps {
     playHandler: (timeOverride?: number) => void;
